feat(separated LoadedReusultInstanceManager script)
