+++ Assets/Scripts/Frontend/LoadSavedResult.cs
@@ -187,28 +187,3 @@ public class LoadSavedResult : MonoBehaviour
     
 }
 
-
-//Script for donot destroy game object script for  persistent junction data 
-
-public class LoadedReusultInstanceManager : MonoBehaviour
-{
-    public static LoadedReusultInstanceManager Instance; 
-
-    public ResultJunctionEntrance northResult; 
-    public ResultJunctionEntrance southResult; 
-    public ResultJunctionEntrance eastResult; 
-    public ResultJunctionEntrance westResult; 
-
-    
-    private void Awake()
-    {
-        if (Instance != null)
-        {
-            Destroy(gameObject);
-            return;
-        }
-
-        Instance = this;
-        DontDestroyOnLoad(gameObject);
-    }
-}
