+++ Assets/Scripts/Frontend/LoadedReusultInstanceManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 13c1c8fe5e4384a4cb90ef38e5ebb16c
